Changes esolved feedback comments
