+++ FFCDemoPaymentService.Tests/TestResults/coverage.info
@@ -168,22 +168,22 @@ FNH:0
 end_of_record
 SF:/c/Repos/ffc-demo-payment-service-core/FFCDemoPaymentService/Messaging/BrokerUrl.cs
 FN:22,System.String FFCDemoPaymentService.Messaging.BrokerUrl::ToString()
-FNDA:4,System.String FFCDemoPaymentService.Messaging.BrokerUrl::ToString()
-DA:23,4
-DA:24,4
-DA:25,4
+FNDA:2,System.String FFCDemoPaymentService.Messaging.BrokerUrl::ToString()
+DA:23,2
+DA:24,2
+DA:25,2
 BRDA:24,20,0,1
-BRDA:24,20,1,3
+BRDA:24,20,1,1
 FN:12,System.Void FFCDemoPaymentService.Messaging.BrokerUrl::.ctor(System.String,System.Int32,System.String,System.String,System.Boolean)
-FNDA:4,System.Void FFCDemoPaymentService.Messaging.BrokerUrl::.ctor(System.String,System.Int32,System.String,System.String,System.Boolean)
-DA:13,4
-DA:14,4
-DA:15,4
-DA:16,4
-DA:17,4
-DA:18,4
-DA:19,4
-DA:20,4
+FNDA:2,System.Void FFCDemoPaymentService.Messaging.BrokerUrl::.ctor(System.String,System.Int32,System.String,System.String,System.Boolean)
+DA:13,2
+DA:14,2
+DA:15,2
+DA:16,2
+DA:17,2
+DA:18,2
+DA:19,2
+DA:20,2
 LF:11
 LH:11
 BRF:2
@@ -199,13 +199,13 @@ DA:25,0
 DA:26,0
 DA:27,0
 FN:29,System.Void FFCDemoPaymentService.Messaging.MessageService::Listen()
-FNDA:2,System.Void FFCDemoPaymentService.Messaging.MessageService::Listen()
-DA:30,2
-DA:31,2
-DA:32,2
-DA:33,2
+FNDA:0,System.Void FFCDemoPaymentService.Messaging.MessageService::Listen()
+DA:30,0
+DA:31,0
+DA:32,0
+DA:33,0
 DA:34,0
-DA:35,2
+DA:35,0
 DA:36,0
 DA:37,0
 DA:38,0
@@ -218,59 +218,59 @@ DA:44,0
 DA:45,0
 DA:46,0
 FN:48,System.Void FFCDemoPaymentService.Messaging.MessageService::CreateConnectionToQueue()
-FNDA:2,System.Void FFCDemoPaymentService.Messaging.MessageService::CreateConnectionToQueue()
-DA:49,2
-DA:50,2
-DA:51,4
-DA:52,4
-DA:53,4
-DA:54,2
-DA:55,2
+FNDA:0,System.Void FFCDemoPaymentService.Messaging.MessageService::CreateConnectionToQueue()
+DA:49,0
+DA:50,0
+DA:51,0
+DA:52,0
+DA:53,0
+DA:54,0
+DA:55,0
 FN:57,System.Void FFCDemoPaymentService.Messaging.MessageService::ReceiveMessage(Amqp.Message)
 FNDA:0,System.Void FFCDemoPaymentService.Messaging.MessageService::ReceiveMessage(Amqp.Message)
 DA:58,0
 DA:59,0
 DA:60,0
 FN:15,System.Void FFCDemoPaymentService.Messaging.MessageService::.ctor(FFCDemoPaymentService.Messaging.IConnection,FFCDemoPaymentService.Messaging.MessageConfig,Microsoft.Extensions.DependencyInjection.IServiceScopeFactory)
-FNDA:10,System.Void FFCDemoPaymentService.Messaging.MessageService::.ctor(FFCDemoPaymentService.Messaging.IConnection,FFCDemoPaymentService.Messaging.MessageConfig,Microsoft.Extensions.DependencyInjection.IServiceScopeFactory)
-DA:16,10
-DA:17,10
-DA:18,10
-DA:19,10
-DA:20,10
-DA:21,10
+FNDA:0,System.Void FFCDemoPaymentService.Messaging.MessageService::.ctor(FFCDemoPaymentService.Messaging.IConnection,FFCDemoPaymentService.Messaging.MessageConfig,Microsoft.Extensions.DependencyInjection.IServiceScopeFactory)
+DA:16,0
+DA:17,0
+DA:18,0
+DA:19,0
+DA:20,0
+DA:21,0
 LF:37
-LH:18
+LH:0
 BRF:0
 BRH:0
 FNF:5
-FNH:3
+FNH:0
 end_of_record
 SF:/c/Repos/ffc-demo-payment-service-core/FFCDemoPaymentService/Messaging/MesssageConfig.cs
 FN:7,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_Host()
-FNDA:13,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_Host()
-DA:8,13
+FNDA:0,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_Host()
+DA:8,0
 FN:8,System.Int32 FFCDemoPaymentService.Messaging.MessageConfig::get_Port()
-FNDA:13,System.Int32 FFCDemoPaymentService.Messaging.MessageConfig::get_Port()
-DA:9,13
+FNDA:0,System.Int32 FFCDemoPaymentService.Messaging.MessageConfig::get_Port()
+DA:9,0
 FN:9,System.Boolean FFCDemoPaymentService.Messaging.MessageConfig::get_UseSsl()
-FNDA:13,System.Boolean FFCDemoPaymentService.Messaging.MessageConfig::get_UseSsl()
-DA:10,13
+FNDA:0,System.Boolean FFCDemoPaymentService.Messaging.MessageConfig::get_UseSsl()
+DA:10,0
 FN:10,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentQueue()
-FNDA:13,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentQueue()
-DA:11,13
+FNDA:0,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentQueue()
+DA:11,0
 FN:11,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentUserName()
-FNDA:13,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentUserName()
-DA:12,13
+FNDA:0,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentUserName()
+DA:12,0
 FN:12,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentPassword()
-FNDA:13,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentPassword()
-DA:13,13
+FNDA:0,System.String FFCDemoPaymentService.Messaging.MessageConfig::get_PaymentPassword()
+DA:13,0
 LF:6
-LH:6
+LH:0
 BRF:0
 BRH:0
 FNF:6
-FNH:6
+FNH:0
 end_of_record
 SF:/c/Repos/ffc-demo-payment-service-core/FFCDemoPaymentService/Data/ApplicationDbContext.cs
 FN:11,Microsoft.EntityFrameworkCore.DbSet`1<FFCDemoPaymentService.Models.Schedule> FFCDemoPaymentService.Data.ApplicationDbContext::get_Schedule()
